requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
flask
plotly
pandas
boto3
